Implemented login + gameupdate + gamerooms + changed client command handling
Implemented:

Login handling(ServerHandler + ServerXmlHandler):
The server can now recieve a login from the client(usernames and hashed password) and compare it to a locally stored XML file which contains profiles. This XML file also contains the users total win and loss amounts(might make a ranking system if there's time).

GameUpdate:
The simple yet elegant bread and butter of how players moves will be handled by the server. As each instance of Player contains their field, graveyard, health etc each GameUpdate instance will contain a serialized copy of each Player instance of the players, plus the gameID of the game it relates to. The plan is when each player makes their move and updates various aspects of the game and ends their turn, the client will make a GameUpdate and send it to the server to be sent to the other player which will be waiting for it. Upon recieving a GameUpdate the client will update their local GameInstance and then be allowed to play their turn. This required the server to have copies of classes such as the graveyard and field.
It also contains a fromPlayer var which will help keep track of which player sent the GameUpdate, meaning th server can avoid sending that GameUpdate back to the client who sent it.

GameRoom & GameRoomList(+NetworkGameList):
GameRoom will be how the server keeps track of individual games. Each instance of GameRoom will contain the OutputStreams of both players plus a list of any game observers OutputStreams. This will ensure that when a GameUpdate is recieved by the server it know where to send it on to.

NetworkGameRoom was implemented seperate to just send over a list of games to clients without the client having to recieve each instance of GameRoom, this is because the client has no need to have the OutputStreams within the GameRoom.

Command:
The Command class supersedes the old Message class in functionality. It still works in a similar way, with the commands being text that is parsed by the server, but it now also contains an optional sentFrom var to track what Player profile the command was sent from.